refactor: remove commons-math3 and commons-text dependencies
Replace RandomDataGenerator with Java 17 Random API (nextInt/nextLong/nextDouble
with origin/bound). Replace StringEscapeUtils with TextUtils.escapeNonAscii().
Removes ~2.5MB of runtime dependencies.

+++ src/main/java/com/devskiller/jfairy/producer/BaseProducer.java
@@ -117,7 +117,7 @@ public class BaseProducer {
 	 * @return pseudorandom {@code long} value between {@code mon} (inclusive) and {@code max} (inclusive)
 	 */
 	public long randomBetween(long min, long max) {
-		return random.nextDouble(min, max);
+		return random.nextLong(min, max);
 	}
 
 	/**
